+++ packages/runtime/test/__modules__/test/index.js
@@ -1,0 +1,3 @@
+export const x = 'test';
+
+export default x;